add scrit site_5.py multi pages and clear functions
