+++ .gitignore
@@ -4,6 +4,7 @@ maisons*.yaml
 appartements*.yaml
 logements.yaml
 __pycache__
+functions/__pycache__
 .env
 notes.md
 site5.txt
