+++ readme.md
@@ -1,4 +1,3 @@
-
 # Property Ad Watch
 
 Property Ad Watch is a Python-based web scraping project that collects property ads from multiple websites. The project centralizes the collected data in a shared YAML file for easy analysis and storage.
@@ -25,11 +24,8 @@ property_adwatch/
 ├── functions/
 |   ├── _init_.py
 |   ├── email_utils.py
-|   └── scrapping.py
-│
-├── scripts/
-|   ├── extract_features.py     # Utility to save data to a shared YAML file
-|   └── extract_number_from_type.py
+|   ├── scrapping.py
+|   └── scripts.py
 |
 ├── main.py             # main python script to run all specific python scripts
 ├── site_0.py           # Scraper for site 0
@@ -53,7 +49,7 @@ property_adwatch/
 - The main entry point for running all scrapers is main.py.
 - GitHub Actions is configured to execute the scripts automatically on:
     - Push to the main branch.
-    - Scheduled daily at 8:00 AM UTC.
+    - Scheduled daily at 7:00 AM UTC.
     - Manual trigger on other branches.
 
 ##  Usage
@@ -66,7 +62,30 @@ property_adwatch/
 pip install -r requirements.txt
 ```
 #### Run Locally
-To scrape data manually, execute the scraper scripts with the desired URL:
+To scrape data manually
+
+1. create .env file in the main repository directory with the following values : 
+
+```sh
+URL_BAILLEUR_0=""
+URL_BAILLEUR_1=""
+URL_BAILLEUR_2=""
+URL_BAILLEUR_4=""
+URL_BAILLEUR_5=""
+
+# URL_BAILLEUR_3="" not working yet
+ 
+EMAIL_SENDER="<username>a@gmail.com"
+EMAIL_RECIPIENTS=["username@gmail.com","username@gmail.com"]
+EMAIL_PASSWORD="xxxx xxxx xxxx xxxx"
+EMAIL_THREAD_ID_MAISONS="<xxxxxxxxxxxxxxxxxx@mx.google.com>"
+EMAIL_THREAD_ID_APPARTEMENTS="xxxxxxxxxxxxxx@mx.google.com>"
+```
+
+Notes: 
+    -   you can get create your gmail app password here https://myaccount.google.com/apppasswords 
+
+2. execute the scraper scripts with the desired URL:
 
 ```sh
 python <MY_SITE>.py "https://<MY_SITE>"
